+++ ext_emconf.php
@@ -11,7 +11,7 @@ $EM_CONF[$_EXTKEY] = [
             'typo3' => '11.5.17-11.5.99',
             'extbase' => '11.5.17-11.5.99',
             'fluid' => '11.5.17-11.5.99',
-            'feuserextrafields' => '5.3.0-6.99.99',
+            'feuserextrafields' => '6.2.0-6.99.99',
             'oelib' => '5.1.0-6.99.99',
         ],
         'conflicts' => [