+++ rector.php
@@ -97,7 +97,7 @@ return RectorConfig::configure()
     ->withImportNames(true, true, false)
     ->withConfiguredRule(ExtEmConfRector::class, [
         ExtEmConfRector::PHP_VERSION_CONSTRAINT => '7.4.0-8.3.99',
-        ExtEmConfRector::TYPO3_VERSION_CONSTRAINT => '11.5.17-11.5.99',
+        ExtEmConfRector::TYPO3_VERSION_CONSTRAINT => '11.5.37-11.5.99',
         ExtEmConfRector::ADDITIONAL_VALUES_TO_BE_REMOVED => [],
     ])
     ->withSkip([