+++ CHANGELOG.md
@@ -14,6 +14,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 ### Changed
 
 ### Removed
+- Drop support for PHP 7.2/7.3 (#796)
 
 ### Fixed