+++ CHANGELOG.md
@@ -22,6 +22,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 - Drop support for PHP 7.2/7.3 (#796)
 
 ### Fixed
+- Drop obsolete entries from the `ext_emconf.php` (#860)
 
 ## 6.4.0