+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 ## x.y.z
 
 ### Added
+- Add a field for the VAT IN (#881)
 - Record the date of acceptance for the terms and privacy (#880)
 - Add a checkbox for the terms & conditions (#877, #879)