+++ CHANGELOG.md
@@ -13,6 +13,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 ### Deprecated
 
 ### Changed
+- Require TYPO3 >= 11.5.37 (#949)
 - Switch to constructor injection (#864, #865)
 - Use PHP 7.4 language features and add native type declarations (#817)
 - Require `feuserextrafields` >= 6.3.0 (#805, #873)